+++ packages/util/src/bi/toU8a.spec.ts
@@ -4,12 +4,29 @@
 /// <reference types="@polkadot/dev-test/globals.d.ts" />
 
 import { arrayRange } from '../array/index.js';
+import { TESTS } from '../bn/toU8a.spec.js';
 import { perf } from '../test/index.js';
 import { nToU8a } from './index.js';
 
 const ptest = arrayRange(65536).map((v) => [v]);
 
 describe('nToU8a', (): void => {
+  describe('conversion tests', (): void => {
+    for (let i = 0, count = TESTS.length; i < count; i++) {
+      const [isLe, isNegative, numarr, strval] = TESTS[i];
+      const bitLength = numarr.length * 8;
+
+      it(`${i}: converts from ${strval} (bitLength=${bitLength}, isLe=${isLe}, isNegative=${isNegative})`, (): void => {
+        expect(
+          nToU8a(
+            BigInt(strval),
+            { bitLength, isLe, isNegative }
+          )
+        ).toEqual(new Uint8Array(numarr));
+      });
+    }
+  });
+
   it('converts null values to 0x00', (): void => {
     expect(
       nToU8a(null)
@@ -38,24 +55,6 @@ describe('nToU8a', (): void => {
     expect(
       nToU8a(0x123456n, { bitLength: 32, isLe: true })
     ).toEqual(new Uint8Array([0x56, 0x34, 0x12, 0x00]));
-  });
-
-  it('converts negative numbers', (): void => {
-    expect(
-      nToU8a(-1234n, { isNegative: true })
-    ).toEqual(new Uint8Array([46, 251]));
-  });
-
-  it('converts negative numbers (BE)', (): void => {
-    expect(
-      nToU8a(-1234n, { isLe: false, isNegative: true })
-    ).toEqual(new Uint8Array([251, 46]));
-  });
-
-  it('converts negative numbers (bitLength)', (): void => {
-    expect(
-      nToU8a(-1234n, { bitLength: 32, isNegative: true })
-    ).toEqual(new Uint8Array([46, 251, 255, 255]));
   });
 
   perf('nToU8a', 250000, ptest, nToU8a);